# backend/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  piemdm_internal_model.FieldAttachment:
    properties:
      accept:
        description: 允许的文件类型
        items:
          type: string
        type: array
      maxSize:
        description: 最大文件大小（字节）
        type: integer
      multiple:
        description: 是否支持多文件上传
        type: boolean
    type: object
  piemdm_internal_model.FieldBehavior:
    properties:
      defaultValue: {}
      readOnly:
        type: boolean
      show:
        description: 是否显示 (对应 is_show)
        type: boolean
      visibleOn:
        description: 表达式
        type: string
    type: object
  piemdm_internal_model.FieldDataSource:
    properties:
      api:
        description: API
        type: string
      fillBack:
        additionalProperties:
          type: string
        type: object
      filter:
        additionalProperties: {}
        type: object
      foreignKey:
        type: string
      labelField:
        type: string
      options:
        description: Static
        items:
          $ref: '#/definitions/piemdm_internal_model.OptionItem'
        type: array
      targetTable:
        description: Relation
        type: string
      type:
        description: static, api, relation
        type: string
      valueField:
        type: string
    type: object
  piemdm_internal_model.FieldDateTime:
    properties:
      dateFormat:
        description: 日期格式
        type: string
      defaultToCurrentTime:
        description: 创建时默认当前时间
        type: boolean
      onUpdateToCurrentTime:
        description: 更新时自动更新为当前时间
        type: boolean
      showTime:
        description: 是否显示时间选择器
        type: boolean
      timeFormat:
        description: 时间格式
        type: string
      timezone:
        description: 是否支持时区
        type: boolean
    type: object
  piemdm_internal_model.FieldOptions:
    properties:
      attachment:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldAttachment'
        description: 附件
      behavior:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldBehavior'
        description: 行为配置
      datasource:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldDataSource'
        description: 数据源 (含关联) - 保留兼容
      datetime:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldDateTime'
        description: 日期时间行为
      patterns:
        description: 自动编码
        items:
          $ref: '#/definitions/piemdm_internal_model.SequencePattern'
        type: array
      relation:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldRelation'
        description: 新增配置
      richtext:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldRichText'
        description: 富文本
      sort:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldSort'
        description: 排序字段
      trim:
        description: 字符串去空格
        type: boolean
      ui:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldUI'
        description: 前端 UI 配置
      validation:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldValidation'
        description: 验证规则
    type: object
  piemdm_internal_model.FieldRelation:
    properties:
      filter:
        additionalProperties: {}
        description: 过滤条件
        type: object
      foreignKey:
        description: 外键（可自动生成）
        type: string
      labelField:
        description: 显示字段（默认 name）
        type: string
      multiple:
        description: 是否多选
        type: boolean
      target:
        description: 关联表
        type: string
      valueField:
        description: 存储字段 = 关联键（默认 code）
        type: string
    type: object
  piemdm_internal_model.FieldRichText:
    properties:
      fileCollection:
        description: 附件存储集合名称
        type: string
      length:
        description: short, long
        type: string
      toolbar:
        description: 工具栏配置
        items:
          type: string
        type: array
    type: object
  piemdm_internal_model.FieldSort:
    properties:
      scopeKey:
        description: 排序作用域字段（如按 status 分组排序）
        type: string
    type: object
  piemdm_internal_model.FieldUI:
    properties:
      placeholder:
        description: 占位符 (可选)
        type: string
      title:
        description: 标题 (可选)
        type: string
      widget:
        description: 组件类型 (必填)
        type: string
      widgetProps:
        additionalProperties: {}
        description: 组件参数 (可选)
        type: object
    type: object
  piemdm_internal_model.FieldValidation:
    properties:
      format:
        description: '格式: phone, email, url'
        type: string
      max:
        description: 最大值
        type: integer
      message:
        description: 自定义错误消息
        type: string
      min:
        description: 最小值
        type: integer
      pattern:
        description: 自定义正则
        type: string
      precision:
        description: 总位数 (仅 decimal 类型)
        type: integer
      required:
        type: boolean
      scale:
        description: 小数位数 (仅 decimal 类型)
        type: integer
      validator:
        description: '预定义验证器: integer, email, phone, url'
        type: string
    type: object
...
  piemdm_internal_model.OptionItem:
    properties:
      color:
        type: string
      label:
        type: string
      value: {}
    type: object
...
  piemdm_internal_model.SequencePattern:
    properties:
      options:
        additionalProperties: {}
        description: 配置选项
        type: object
      type:
        description: string, integer, date
        type: string
    type: object
...
  piemdm_internal_model.TableField:
    properties:
      code:
        description: 编码
        maxLength: 64
        type: string
      createdAt:
        type: string
      createdBy:
        description: 创建人
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 属性描述
        maxLength: 256
        type: string
      fieldType:
        description: |-
          FieldType 目前会保存字段格式，在后端进行字段默认配置，字段校验等
          后端的 FieldPreset 预设系统设计得非常完善，每个 FieldType 已经包含：
            DataType (数据库类型)
            Length (字段长度)
            UI.Widget (UI 组件)
            Validation (验证规则)
          结论：不需要移除 FieldType，它是一个很好的设计！
        maxLength: 32
        type: string
      groupName:
        description: 组名
        maxLength: 64
        type: string
      id:
        description: ===== field attr
        type: integer
      indexName:
        description: 索引名称，扩展视图使用，不对外显示
        maxLength: 128
        type: string
      indexPriority:
        description: 索引名称，扩展视图使用，不对外显示
        maximum: 3
        type: integer
      isFilter:
        description: ===== show style
        maxLength: 8
        type: string
      isIndex:
        description: 是否创建索引，Yes,No
        maxLength: 8
        type: string
      isShow:
        description: 是否展示字段  Yes：是  No-否
        maxLength: 8
        type: string
      isUnique:
        description: 唯一标识，Yes,No
        maxLength: 8
        type: string
      length:
        description: 属性长度
        type: integer
      name:
        description: 属性名称
        maxLength: 128
        type: string
      options:
        allOf:
        - $ref: '#/definitions/piemdm_internal_model.FieldOptions'
        description: ========== Options JSON 字段
      required:
        description: 是否必填 Yes：必填 No:不必填
        maxLength: 8
        type: string
      sort:
        description: 显示顺序
        type: integer
      status:
        description: |-
          ========== 状态
          状态：Normal 正常 Frozen 已冻结 Deleted 已删除
        type: string
      tableCode:
        description: 所属模型
        maxLength: 64
        type: string
      type:
        description: '数据库字段类型(DB层): Text, Integer, Date, Decimal 等'
        maxLength: 16
        type: string
      updatedAt:
        type: string
      updatedBy:
        description: 更新人
        type: string
    required:
    - code
    - length
    - name
    - tableCode
    - type
    type: object
...
host: localhost:8787
info:
  contact:
    email: support@piemdm.com
    name: API Support
  description: PieMDM Master Data Management System API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PieMDM API
  version: "1.0"
paths:
...
  /admin/table/{table_code}/options:
    get:
      consumes:
      - application/json
      parameters:
      - description: 表编码
        in: path
        name: table_code
        required: true
        type: string
      - description: 过滤条件(JSON格式)
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      summary: 获取表的选项列表
      tags:
      - 表字段管理
...
  /admin/table_fields:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 15
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 表编码
        in: query
        name: table_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.TableField'
            type: array
      summary: 获取表字段列表
      tags:
      - 表字段管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.TableField'
      summary: 创建表字段
      tags:
      - 表字段管理
  /admin/table_fields/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 字段ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.TableField'
      summary: 获取表字段详情
      tags:
      - 表字段管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 字段ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新表字段
      tags:
      - 表字段管理
  /admin/table_fields/batch:
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量更新表字段状态
      tags:
      - 表字段管理
  /admin/table_fields/batch_delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 批量删除请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量删除表字段
      tags:
      - 表字段管理
  /admin/table_fields/fields:
    get:
      consumes:
      - application/json
      parameters:
      - description: 表编码
        in: query
        name: table_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.TableField'
            type: array
      summary: 获取表的所有字段
      tags:
      - 表字段管理
  /admin/table_fields/public:
    post:
      consumes:
      - application/json
      parameters:
      - description: 发布请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 发布表结构
      tags:
      - 表字段管理
  /admin/table_permissions:
    get:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: query
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      summary: 获取角色的表权限列表
      tags:
      - 表权限管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 权限信息
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 授予表权限
      tags:
      - 表权限管理
  /admin/table_permissions/batch:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量删除表权限
      tags:
      - 表权限管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量更新表权限状态
      tags:
      - 表权限管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"